Add a Thai keyboard without prediction support
This commit add keyboard layouts in Thai language, along with an empty
language plugin and language features. This language differs enough from
most western languages that it cannot use the common westernsupport
base.

It's easy to add keyboard layouts. It's a different story to add a
decent word prediction. I would need to find an example text for it to
create the n-gram database and would have to write a language features
more or less from scratch. Besides, the text2ngram utility seems to
assume that words are separated by whitespaces, which isn't the case
for Thai. So, let's get keyboard layouts out; at least you will not
encouter bad autocorrection (!).